Return to "Flyspell" for spelling.
Wucuo didn't have a prog-mode option, this led to multiple variables and
functions to be considered as misspelling. For this reason, it was decided to
return to Flyspell, and define a function to determine when a word should be
ignored depending if its misspelled and if its not in camel-case or pascal-case.

This function is not enabled by default and it should be specific to major modes.